Added support for 1.21 - 1.21.5

+++ src/main/java/net/kakos1220/titlescreencleaner/mixin/SplashTextRendererMixin.java
@@ -13,7 +13,7 @@ public class SplashTextRendererMixin {
      * @reason Removes SplashText rendering entirely
      */
     @Overwrite
-    public void render(DrawContext context, int screenWidth, TextRenderer textRenderer, float alpha) {
+    public void render(DrawContext context, int screenWidth, TextRenderer textRenderer, int alpha) {
 
     }
 }
